@@ -38,6 +38,14 @@
     "string-math": "^1.2.1"
   },
   "devDependencies": {
-    "eslint": "^6.3.0"
+    "@types/atob": "^2.1.2",
+    "@types/btoa": "^1.2.3",
+    "@types/dateformat": "^3.0.0",
+    "@types/node": "^12.7.5",
+    "@types/pg": "^7.11.1",
+    "@typescript-eslint/eslint-plugin": "^2.2.0",
+    "@typescript-eslint/parser": "^2.2.0",
+    "eslint": "^6.4.0",
+    "typescript": "^3.6.3"
   }
 }
