@@ -1,12 +1,15 @@
-const atob = require('atob');
-const btoa = require('btoa');
+import ReknownClient from '../../structures/client';
+import { Message, TextChannel, MessageEmbed } from 'discord.js';
+import atob from 'atob';
+import btoa from 'btoa';
 
-module.exports.run = (client, message, args) => {
-  if (!message.channel.permissionsFor(client.user).has('EMBED_LINKS')) return client.functions.noClientPerms(message, [ 'Embed Links' ], message.channel);
+module.exports.run = (client: ReknownClient, message: Message, args: string[]): void => {
+  // eslint-disable-next-line no-extra-parens
+  if (!(message.channel as TextChannel).permissionsFor(client.user).has('EMBED_LINKS')) return client.functions.noClientPerms(message, [ 'Embed Links' ], message.channel);
 
   const method = args[1] ? args[1].toLowerCase() : null;
   if (!method) return client.functions.noArg(message, 1, 'an action to do, either decode / encode.');
-  if (method !== 'decode' && method !== 'encode') return client.functions.badArg.run(message, 1, 'The action provided was not a valid choice. It must be either decode or encode.');
+  if (method !== 'decode' && method !== 'encode') return client.functions.badArg(message, 1, 'The action provided was not a valid choice. It must be either decode or encode.');
 
   const input = args.slice(2).join(' ');
   if (!input) return client.functions.noArg(message, 2, 'an input to decode or encode.');
@@ -15,14 +18,14 @@ module.exports.run = (client, message, args) => {
   if (!result) return client.functions.badArg(message, 2, 'The input provided was not a valid Base64 string.');
   if (result.length > 2048) return client.functions.badArg(message, 2, 'The output was longer than 2048 characters, which is more than a message can hold. Please shorten the input.');
 
-  const embed = new client.MessageEmbed()
+  const embed = new MessageEmbed()
     .setColor(client.config.embedColor)
     .setDescription(client.escMD(result))
     .setFooter(`Successfully ${method}d! | Requested by ${message.author.tag}`, message.author.displayAvatarURL())
     .setTimestamp()
     .setTitle('Output');
 
-  return message.channel.send(embed);
+  return void message.channel.send(embed);
 };
 
 module.exports.help = {
