@@ -1,4 +1,7 @@
-module.exports.run = async (client, message, args) => {
+import ReknownClient from '../../structures/client';
+import { Message } from 'discord.js';
+
+module.exports.run = async (client: ReknownClient, message: Message, args: string[]): Promise<void> => {
   if (!message.guild.me.hasPermission('MANAGE_ROLES')) return client.functions.noClientPerms(message, [ 'Manage Roles' ]);
   if (!args[1]) return client.functions.noArg(message, 1, 'a role to mention.');
   const role = message.guild.roles.find(r => r.name === args.slice(1).join(' ')) || message.guild.roles.get(args[1]);
@@ -9,7 +12,7 @@ module.exports.run = async (client, message, args) => {
   if (!mentionable) {
     if ((!message.member.hasPermission('MANAGE_ROLES') ||
     role.position >= message.member.roles.highest.position) &&
-    message.author.id !== message.guild.ownerID) return message.channel.send('You do not have enough permissions to mention that role. You need the `Manage Roles` permission and also have to be above the mentioning role.');
+    message.author.id !== message.guild.ownerID) return void message.channel.send('You do not have enough permissions to mention that role. You need the `Manage Roles` permission and also have to be above the mentioning role.');
 
     await role.setMentionable(true);
   }
